Move `transform_response` to dedicated helpers
